AutoWorkbench should return !0 when doAdd is false
if doAdd is false, the AutoWorkBench now estimates the amount of free space it has, and return that as how many can be added. This resolves MustHave issue #160.
if addItem(doAdd:false) returns 0, some code branches assume that this item can not recieve anything, when it can.

Tested with obsidian pipe connected to workbench, chest next to workbench, hopper above, and general pipes providing. All are delivered as expected.