Fix for quarry stalling when diagonal movement requested
a quarry should ask the arm what the minimum velocity is, and only request a fixed multiple of that. That way excess energy isn't wasted, and there is a chance for the energy to build up and move at 2/32 per tick occasionally.

This non-wasting of energy also restores the behavior of a quarry to continue mining after power input is cut.