a Transactor should be able to request addition to a specific slot
The autocrafting table searches for a specific slot to add an item, but
then fails to actually place it into that specific slot beacuse
Transactor then decides which slot it should go into -- a
add/injectSpecific funcion call has been added which takes a slot and
tries to put the stack in there. THe exception is the TransactorSPecial
which assumes (even on an addSpecifc call) that the ISpecial is the best
entitty to decide which slot it should go into/

Because the AutoWorkbench tries to split stacks, rather than fill them
(eg making ladders from sticks) large stacks should be added one at a
time to distribute them between internal stacks. There is a minor
performance boos when items get added one per time because the stack
doesn't need to be cloned to prevent change.