Buildcraft should now be mostly MFFS safe.
Depreciated the unbreakableBlock() thing in the API and added a new
function to BlockUtils that does a better job of the same thing.

The Quarry and Mining Well should be completely incapable of penetrating
a Force Field.

The Filler will terminate its current job if it detects an unbreackable
block during an operation, but its still theoretically possible to
affect blocks inside a Force Field. Some patterns are worse than others.
Preventing this completely is pretty difficult without some way to know
whether a block is inside a Force Field.