MachineHookRegistry
This is a small commit on an extension to the Buildcraft API. Simply said, it is a simple API that provides callback functions from Buildcraft machines. The main reason I am writing this is because I am planning on writing an Universal Electricity compatibility layer which will required these three call backs from all machines that can accept energy in order to work. I will be adding more stuff to complete this API in the near future.