Enable facade creation for extra blocks
Adds a configuration option allowing the end user to specify a comma delimited
list of block IDs to be added to BuildCraft as facades.

BuildCraft currently cycles through the standard Minecraft blocks and adds
those automatically, but mods which add their own blocks must tell BuildCraft
to create facades. Those mods which do not, lack facades completely. A prime
example of this is RedPower2 with Marble, Basalt and their subtypes.