Randomize potential oil biome spawns on worldseed
Also reduced Oil Ocean Biome chances again.

Previously it was using the same noise field for all seeds. Now it
randomly offsets the noise field based on the world seed. Its not truly
random, but at least you won't be able to travel to the same coordinates
in every world and find an Oil Biome.