Make Flatten less stupid
It now processes in columns, greatly increasing the performance.

It is no longer halted entirely by anything it can't break. In other
words, you can use it to fill over bedrock.

Changed how the clear portion works, it no longer clears all the way to
sky, instead it only clears IN the box.