Fix the refinery stopping when the first tank is almost empty
I found this bug when I was refining biomass into biofuel. I had piping in the system, so it wasn't just full buckets, and somehow it got an amount that isn't a multiple of 4, and the pistons became red and the refinery stopped working completely. I could see that there was a very small amount of biomass still in the first tank, but it was probably not enough. This will make it drain a bit from the second tank if it's got the same fluid.