Changing  serverSideRequired to false
If you have it to true, The Client isn't able to connect to a server without the mod ("
Tells Forge how to handle what happens when the client xor the server has the client installed. Takes two parameters.
clientSideRequired 
Asks if you need this on the client to use this mod. This should be true.
serverSideRequired 
Asks if you need this on the server for the client to be able to connect. This should always be false, else you can't join a server if the server doesn't have the mod installed, but you do. It is false by default anyway."
from http://www.minecraftforge.net/wiki/Basic_Modding