Fixed hundreds of warnings
Most of them we caused by the deprecation of various blueprint-related
stuff. The README says that we should
"only submit "code cleanup", if the changes actually have a substantial
impact on readability." And, in my humble opinion, fixing cca. 500
warnings while not changing how the code works at all is a substantial
impact on readability.