Blueprint serialization over the network
This patch changes a bit the behavior of the Library. Instead of being an
item browsing blueprints on the server, it's going to retreive blueprints
built on the server for the client, and then allow the client to send back
its blueprints to the server.

This patch is a first milestone, with functional serialization and
deserialization of a blueprint over the newtork, using the Library as interface.