Fixes handling of template extensions length != 3.
When a template specifies an extension that is not of length 3 (such as the default `tpl` and `bpt` ones), loading the list currently generates broken entries, because the `uniqueId` is built assuming the extension length is always 3. The result is that such entries can neither be loaded nor deleted. This change should properly allow for arbitrary file extension lengths.