Add a formatter to be used accross the entrety of buildcraft.
Because this 1.8 port has already been reformatted many times away from the original (so that the blame view of github cannot be used for anything other than for me, and that is after the reformatting) it seems sensible to adopt a singular formatter for the entire project. However, this may or may not be a good idea for the future of buildcraft, so I'm asking @asiekierka for your thoughts.

(Also I'm wondering if this formatter is considered "usable" at all, and if there are any objections to any specfics. Its annoying that this is only for eclipse (and i don't know what IDEA does for formatting so that may or may not be a seperate issue)