Move the fluid rebalancing for tanks out to a (public) method.
This implements #4029 (with minor changes to make it work as intended).

Make GuideManager.PAGE_LOADERS public, so that addons can add any IPageLoader themselves. I'm not sure if this is actually useful though, as we might need to mandate file types for later conversion utilities.
(In addition I'd rather keep every addon using the same markdown format that we use)