Improvements to Heat Exchanger code (#4136)
* Removed unused code.

* Fix #4135 - Made Heat Exchanger to receive multiples of 5

This changed processing speed of Heat Exchanger
allowing to process every combination of fluids from the distiller.

Also removed the multiplier non-divisible by 1000 that
was in contradiction to the comment.

* Made Heat Exchanger process with valid multiples

This fixes a todo on the code on which the multiple of the processing
given by the HE size was fixed, causing no smaller multiples to
be processed.

Tested ingame that 375 mB is now fully processed by the exchanger
in parts (37*10 + 5*1).

* Made Heat Exchanger process arbitrary amounts of fluid.

* Added entry to changelog.

(cherry picked from commit b952a7d)